+++ packages/react-native/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Next
+
+* Add `<Navigation>` component, which lets the user know when asynchronous routes are navigating and cancel the navigation.
+
 ## 1.0.2
 
 * Guard `setState()` calls for unmounted `<Link>`.